+++ client-gwt/src/main/java/org/realityforge/replicant/client/gwt/GwtWebPollerDataLoaderService.java
@@ -174,6 +174,11 @@ public abstract class GwtWebPollerDataLoaderService<T extends ClientSession<T, G
   {
     final RequestBuilder rb = new RequestBuilder( method, url );
     rb.setHeader( "Pragma", "no-cache" );
+    final String authenticationToken = getSessionContext().getAuthenticationToken();
+    if ( null != authenticationToken )
+    {
+      rb.setHeader( "Authorization", "Bearer " + authenticationToken );
+    }
     return rb;
   }
 }